CarBTImprovements
Car now works wirelessly using bluetooth. Work still needs to be done on the steering algorithms, and there are random  interferences of unknown origin to the communications from the controller. It also doesn't work unless there is a delay of at lest 150ms between instructions from the controller for some reason. MichaelCarScript and MichaelControllerScript are the ones that work